Explicitly unmount partitions after copying /boot
When this was left to universal_cleanup there appeared to be a race
condition where boot would still be busy when attempting to unmount it.
The unmount would therefore fail.  This led to the nbd being
disconnected without the writes to the boot partition being committed,
resulting in an empty /boot in the resulting image.